Finaliser l'authontification

+++ var/cache/dev/twig/fa/fa44fe51a62714f56e27b0b1f1352ca409164d87a77a723548db65f93f2161e0.php
@@ -68,20 +68,35 @@ class __TwigTemplate_561837bade794a4a029c33662ab57081ef38a7899e7d542ad77317e1029
     <button class=\"navbar-toggler\" type=\"button\" data-toggle=\"collapse\" data-target=\"#navbarSupportedContent\" aria-controls=\"navbarSupportedContent\" aria-expanded=\"false\" aria-label=\"Toggle navigation\">
         <span class=\"navbar-toggler-icon\"></span>
     </button>
-    <a class=\"btn btn-outline-light\" href=\"";
+    ";
         // line 18
-        echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("Groupes");
-        echo "\"> Mes Groupes   </a>
-
+        if (twig_get_attribute($this->env, $this->source, (isset($context["app"]) || array_key_exists("app", $context) ? $context["app"] : (function () { throw new RuntimeError('Variable "app" does not exist.', 18, $this->source); })()), "user", [], "any", false, false, false, 18)) {
+            // line 19
+            echo "    <a class=\"btn btn-outline-light\" href=\"";
+            echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("Groupes");
+            echo "\"> Mes Groupes   </a>
+    ";
+        }
+        // line 21
+        echo "
 
     <div class=\"collapse navbar-collapse\" id=\"navbarSupportedContent\">
         <ul class=\"navbar-nav mr-auto\">
-            <li class=\"nav-item\">
-                <a class=\"btn btn-outline-light\" href=\"";
-        // line 24
-        echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("Commande");
-        echo "\"> Commande </a>
-            </li>
+            ";
+        // line 25
+        if (twig_get_attribute($this->env, $this->source, (isset($context["app"]) || array_key_exists("app", $context) ? $context["app"] : (function () { throw new RuntimeError('Variable "app" does not exist.', 25, $this->source); })()), "user", [], "any", false, false, false, 25)) {
+            // line 26
+            echo "                <li class=\"nav-item\">
+                    <a class=\"btn btn-outline-light\" href=\"";
+            // line 27
+            echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("Profil");
+            echo "\"> Mon Profil </a>
+                </li>
+             ";
+        }
+        // line 30
+        echo "
+
             <li class=\"nav-item dropdown\">
                 <a class=\"nav-link dropdown-toggle\" href=\"#\" id=\"navbarDropdown\" role=\"button\" data-toggle=\"dropdown\" aria-haspopup=\"true\" aria-expanded=\"false\">
                     Services
@@ -101,38 +116,50 @@ class __TwigTemplate_561837bade794a4a029c33662ab57081ef38a7899e7d542ad77317e1029
         </ul>
         <form class=\"form-inline my-2 my-lg-0\">
             <input class=\"form-control mr-sm-2\" type=\"search\" placeholder=\"Entrez votre recherche\" aria-label=\"Search\">
-            <button class=\"btn btn-outline-light my-2 my-sm-0\" type=\"submit\">Rechercher</button>
+            <button class=\"btn btn-outline-light my-2 my-sm-0\" type=\"submit\" >Rechercher</button>
         </form>
 
 
     </div>
     <li class=\"nav-item\">
-        <a class=\"btn btn-outline-light\" href=\"";
-        // line 51
-        echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("Panier");
-        echo "\"> Panier </a>
-        <a class=\"btn btn-outline-light\" href=\"";
-        // line 52
-        echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("users_new");
-        echo "\"> Inscription
-        </a>
-        <a class=\"btn btn-outline-light\" href=\"";
-        // line 54
-        echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("login");
-        echo "\"> Connexion </a>
-        <a class=\"btn btn-outline-light\" href=\"";
-        // line 55
-        echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("issam");
-        echo "\"> Issam </a>
+        ";
+        // line 57
+        if (twig_get_attribute($this->env, $this->source, (isset($context["app"]) || array_key_exists("app", $context) ? $context["app"] : (function () { throw new RuntimeError('Variable "app" does not exist.', 57, $this->source); })()), "user", [], "any", false, false, false, 57)) {
+            // line 58
+            echo "            <a class=\"btn btn-outline-light\" href=\"";
+            echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("Panier");
+            echo "\"> Mon Panier </a>
+            <a class=\"btn btn-outline-light\" href=\"";
+            // line 59
+            echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("logout");
+            echo "\"> Se déconnecter </a>
+            <a class=\"btn btn-outline-light\" href=\"";
+            // line 60
+            echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("issam");
+            echo "\"> Admin </a>
+        ";
+        } else {
+            // line 62
+            echo "            <a class=\"btn btn-outline-light\" href=\"";
+            echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("users_new");
+            echo "\"> Inscription </a>
+            <a class=\"btn btn-outline-light\" href=\"";
+            // line 63
+            echo $this->extensions['Symfony\Bridge\Twig\Extension\RoutingExtension']->getPath("login");
+            echo "\"> Connexion </a>
+        ";
+        }
+        // line 65
+        echo "
     </li>
 </nav>
 
 ";
-        // line 59
+        // line 69
         $this->displayBlock('body', $context, $blocks);
-        // line 60
+        // line 70
         $this->displayBlock('javascripts', $context, $blocks);
-        // line 61
+        // line 71
         echo "<script src=\"https://code.jquery.com/jquery-3.3.1.slim.min.js\" integrity=\"sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo\" crossorigin=\"anonymous\"></script>
 <script src=\"https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js\" integrity=\"sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1\" crossorigin=\"anonymous\"></script>
 <script src=\"https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js\" integrity=\"sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM\" crossorigin=\"anonymous\"></script>
@@ -184,7 +211,7 @@ class __TwigTemplate_561837bade794a4a029c33662ab57081ef38a7899e7d542ad77317e1029
 
     }
 
-    // line 59
+    // line 69
     public function block_body($context, array $blocks = [])
     {
         $__internal_085b0142806202599c7fe3b329164a92397d8978207a37e79d70b8c52599e33e = $this->extensions["Symfony\\Bundle\\WebProfilerBundle\\Twig\\WebProfilerExtension"];
@@ -201,7 +228,7 @@ class __TwigTemplate_561837bade794a4a029c33662ab57081ef38a7899e7d542ad77317e1029
 
     }
 
-    // line 60
+    // line 70
     public function block_javascripts($context, array $blocks = [])
     {
         $__internal_085b0142806202599c7fe3b329164a92397d8978207a37e79d70b8c52599e33e = $this->extensions["Symfony\\Bundle\\WebProfilerBundle\\Twig\\WebProfilerExtension"];
@@ -230,7 +257,7 @@ class __TwigTemplate_561837bade794a4a029c33662ab57081ef38a7899e7d542ad77317e1029
 
     public function getDebugInfo()
     {
-        return array (  205 => 60,  188 => 59,  171 => 10,  153 => 5,  136 => 61,  134 => 60,  132 => 59,  125 => 55,  121 => 54,  116 => 52,  112 => 51,  82 => 24,  73 => 18,  66 => 14,  61 => 11,  59 => 10,  51 => 5,  45 => 1,);
+        return array (  232 => 70,  215 => 69,  198 => 10,  180 => 5,  163 => 71,  161 => 70,  159 => 69,  153 => 65,  148 => 63,  143 => 62,  138 => 60,  134 => 59,  129 => 58,  127 => 57,  98 => 30,  92 => 27,  89 => 26,  87 => 25,  81 => 21,  75 => 19,  73 => 18,  66 => 14,  61 => 11,  59 => 10,  51 => 5,  45 => 1,);
     }
 
     public function getSourceContext()
@@ -252,14 +279,20 @@ class __TwigTemplate_561837bade794a4a029c33662ab57081ef38a7899e7d542ad77317e1029
     <button class=\"navbar-toggler\" type=\"button\" data-toggle=\"collapse\" data-target=\"#navbarSupportedContent\" aria-controls=\"navbarSupportedContent\" aria-expanded=\"false\" aria-label=\"Toggle navigation\">
         <span class=\"navbar-toggler-icon\"></span>
     </button>
+    {% if app.user %}
     <a class=\"btn btn-outline-light\" href=\"{{ path('Groupes') }}\"> Mes Groupes   </a>
+    {% endif %}
 
 
     <div class=\"collapse navbar-collapse\" id=\"navbarSupportedContent\">
         <ul class=\"navbar-nav mr-auto\">
-            <li class=\"nav-item\">
-                <a class=\"btn btn-outline-light\" href=\"{{ path('Commande') }}\"> Commande </a>
-            </li>
+            {% if app.user %}
+                <li class=\"nav-item\">
+                    <a class=\"btn btn-outline-light\" href=\"{{ path('Profil') }}\"> Mon Profil </a>
+                </li>
+             {% endif %}
+
+
             <li class=\"nav-item dropdown\">
                 <a class=\"nav-link dropdown-toggle\" href=\"#\" id=\"navbarDropdown\" role=\"button\" data-toggle=\"dropdown\" aria-haspopup=\"true\" aria-expanded=\"false\">
                     Services
@@ -279,17 +312,21 @@ class __TwigTemplate_561837bade794a4a029c33662ab57081ef38a7899e7d542ad77317e1029
         </ul>
         <form class=\"form-inline my-2 my-lg-0\">
             <input class=\"form-control mr-sm-2\" type=\"search\" placeholder=\"Entrez votre recherche\" aria-label=\"Search\">
-            <button class=\"btn btn-outline-light my-2 my-sm-0\" type=\"submit\">Rechercher</button>
+            <button class=\"btn btn-outline-light my-2 my-sm-0\" type=\"submit\" >Rechercher</button>
         </form>
 
 
     </div>
     <li class=\"nav-item\">
-        <a class=\"btn btn-outline-light\" href=\"{{ path('Panier') }}\"> Panier </a>
-        <a class=\"btn btn-outline-light\" href=\"{{ path('users_new') }}\"> Inscription
-        </a>
-        <a class=\"btn btn-outline-light\" href=\"{{ path('login') }}\"> Connexion </a>
-        <a class=\"btn btn-outline-light\" href=\"{{ path('issam') }}\"> Issam </a>
+        {% if app.user %}
+            <a class=\"btn btn-outline-light\" href=\"{{ path('Panier') }}\"> Mon Panier </a>
+            <a class=\"btn btn-outline-light\" href=\"{{ path('logout') }}\"> Se déconnecter </a>
+            <a class=\"btn btn-outline-light\" href=\"{{ path('issam') }}\"> Admin </a>
+        {% else %}
+            <a class=\"btn btn-outline-light\" href=\"{{ path('users_new') }}\"> Inscription </a>
+            <a class=\"btn btn-outline-light\" href=\"{{ path('login') }}\"> Connexion </a>
+        {% endif %}
+
     </li>
 </nav>
